restructure side_metadata to a submodule of metadata
