+++ src/util/metadata/side_metadata/global.rs
@@ -1,6 +1,7 @@
 use super::*;
 use crate::util::constants::BYTES_IN_PAGE;
 use crate::util::memory;
+use crate::util::metadata::MetadataSpec;
 use crate::util::{constants, Address};
 use std::sync::atomic::{AtomicU16, AtomicU32, AtomicU8, AtomicUsize, Ordering};
