+++ src/util/metadata/side_metadata/mod.rs
@@ -1,0 +1,15 @@
+mod constants;
+mod helpers;
+#[cfg(target_pointer_width = "32")]
+mod helpers_32;
+
+mod global;
+mod sanity;
+mod side_metadata_tests;
+
+pub use constants::*;
+pub use global::*;
+pub use helpers::*;
+#[cfg(target_pointer_width = "32")]
+pub use helpers_32::*;
+pub use sanity::SideMetadataSanity;
